refactor: updated component implementation for improved structure and functionality
